fmp-approach/requirements.txt
requests==2.25.1
pandas==2.2.3
openpyxl==3.0.7
python-dotenv==0.17.1
numpy==1.26.4
tenacity>=8.2.0 



